:recycle: Move auth route specs into route spec
