+++ src/client/mock_server/routes/mod.rs
@@ -3,3 +3,6 @@ pub mod token;
 
 pub use self::auth::auth_response as auth_route;
 pub use self::token::token as token_route;
+
+#[cfg(test)]
+mod auth_spec;
